moved queueProcessor to service dir
